requirements.txt
asgiref==3.9.2
attrs==25.3.0
coverage==7.10.7
dj-database-url==3.0.1
Django==4.2.24
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.11.0
django-cors-headers==4.6.0
pytest==8.3.3
pytest-django==4.9.0
flake8==7.1.1
requests==2.32.3
